+++ cmd/main.go
@@ -17,11 +17,11 @@ import (
 
 	kitprometheus "github.com/go-kit/kit/metrics/prometheus"
 	"github.com/go-redis/redis"
-	"github.com/mainflux/export/internal/app/export"
-	"github.com/mainflux/export/internal/app/export/api"
-	"github.com/mainflux/export/internal/pkg/messages"
 	"github.com/mainflux/export/pkg/config"
 	exp "github.com/mainflux/export/pkg/config"
+	"github.com/mainflux/export/pkg/export"
+	"github.com/mainflux/export/pkg/export/api"
+	"github.com/mainflux/export/pkg/messages"
 	"github.com/mainflux/mainflux"
 	"github.com/mainflux/mainflux/errors"
 	"github.com/mainflux/mainflux/logger"
